+++ apps/cube_light_and_texture.js
@@ -1,78 +1,84 @@
-/* eslint-disable no-unused-vars */
-
 /**
  * @Author Nikola Lukic
+ * @license MIT
  * @Description Matrix Engine Api Example.
+ * Free for all
  */
-
-/* globals world App world */
 import App from "../program/manifest";
 
 export var runThis = world => {
-	/* globals world App ENUMERATORS SWITCHER OSCILLATOR */
-
 	var textuteImageSamplers = {
 		source: ["res/images/complex_texture_1/diffuse.webp"],
 		mix_operation: "multiply",
 	};
 
+	matrixEngine.Events.camera.yawAmp = 5
 	App.camera.SceneController = true;
 
-	// // Floor
-	world.Add("squareTex", 1, "floor", textuteImageSamplers);
+	// Floor
+	world.Add("cubeLightTex", 1, "floor", textuteImageSamplers);
 	App.scene.floor.position.SetY(0);
-	App.scene.floor.geometry.setScale(20);
-	App.scene.floor.rotation.rotx = -90;
+	App.scene.floor.geometry.setScaleByX(31);
+	App.scene.floor.geometry.setScaleByZ(31);
+	// App.scene.floor.rotation.roty = 90;
 	App.scene.floor.position.y = -1;
-	App.scene.floor.position.z = -20;
+	App.scene.floor.position.z = 0;
+	App.scene.floor.rotation.roty = -90
+	App.scene.floor.setFBO({
+		cameraX: 0,
+		cameraY: 0,
+		cameraZ: 20,
+		pitch: 0,
+		yaw: 0
+	})
+	App.scene.floor.activateShadows('spot-shadow')
 
-	// App.scene.floor.custom.gl_texture = function (object, t) {
-	//   world.GL.gl.bindTexture(world.GL.gl.TEXTURE_2D, object.textures[t]);
-	//   world.GL.gl.texParameteri(
-	//     world.GL.gl.TEXTURE_2D,
-	//     world.GL.gl.TEXTURE_MAG_FILTER,
-	//     world.GL.gl.LINEAR
-	//   );
-	//   world.GL.gl.texParameteri(
-	//     world.GL.gl.TEXTURE_2D,
-	//     world.GL.gl.TEXTURE_MIN_FILTER,
-	//     world.GL.gl.LINEAR
-	//   );
-	//   world.GL.gl.texParameteri(
-	//     world.GL.gl.TEXTURE_2D,
-	//     world.GL.gl.TEXTURE_WRAP_S,
-	//     world.GL.gl.REPEAT
-	//   );
-	//   world.GL.gl.texParameteri(
-	//     world.GL.gl.TEXTURE_2D,
-	//     world.GL.gl.TEXTURE_WRAP_T,
-	//     world.GL.gl.REPEAT
-	//   );
-	//   world.GL.gl.texImage2D(
-	//     world.GL.gl.TEXTURE_2D,
-	//     0,
-	//     world.GL.gl.RGBA,
-	//     world.GL.gl.RGBA,
-	//     world.GL.gl.UNSIGNED_BYTE,
-	//     object.textures[t].image
-	//   );
+	world.Add("squareTex", 1, "MyCubeTex1", textuteImageSamplers);
+	App.scene.MyCubeTex1.activateShadows('spot-shadow')
+	App.scene.MyCubeTex1.geometry.setScaleByX(2);
+	App.scene.MyCubeTex1.geometry.setScaleByY(2);
+	App.scene.MyCubeTex1.position.y = 3;
+	App.scene.MyCubeTex1.position.x = 7;
 
-	//   world.GL.gl.generateMipmap(world.GL.gl.TEXTURE_2D);
-	// };
-
-	world.Add("cubeLightTex", 1, "MyCubeTex1", textuteImageSamplers);
-	world.Add("cubeLightTex", 1, "MyCubeTex2", textuteImageSamplers);
+	setTimeout(() => {
+		App.scene.MyCubeTex1.shadows.lightPosition = [0.1, -0.2, 0.1]
+	}, 200)
 
 
-	App.scene.floor.activateShadows('spot');
+	function onLoadObj(meshes) {
+		for(let key in meshes) {matrixEngine.objLoader.initMeshBuffers(world.GL.gl, meshes[key])}
+		var textuteImageSamplers2 = {
+			source: ["res/images/armor.webp", "res/images/armor.webp"],
+			mix_operation: "multiply",
+		};
+		var textuteImageSamplers = {
+			source: ["res/images/dagger.webp"],
+			mix_operation: "multiply",
+		};
+		world.Add("obj", 1, "armor", textuteImageSamplers2, meshes.armor);
+		App.scene.armor.position.y = 3;
+		App.scene.armor.rotation.rotationSpeed.y = 20;
+		App.scene.armor.LightsData.ambientLight.set(1, 1, 1);
+		App.scene.armor.position.z = -6;
+		App.scene.armor.mesh.setScale(3)
 
-	App.scene.MyCubeTex1.activateShadows('spot-shadow');
+		setTimeout(() => {
+			App.scene.armor.activateShadows('spot-shadow')
+			App.scene.mac.activateShadows('spot-shadow')
 
-	// App.scene.MyCubeTex1.rotation.roty = 45;
-	// App.scene.MyCubeTex2.rotation.roty = -45;
+			App.scene.armor.shadows.lightPosition = [10, 2, -1]
+		}, 200)
 
-	App.scene.MyCubeTex1.position.y = 0;
-	App.scene.MyCubeTex2.position.y = 0;
-	App.scene.MyCubeTex1.position.x = 1;
-	App.scene.MyCubeTex2.position.x = -1;
+		world.Add("obj", 1, "mac", textuteImageSamplers, meshes.mac);
+		App.scene.mac.position.y = 1;
+		App.scene.mac.position.x = -2;
+		App.scene.mac.rotation.rotationSpeed.y = 20;
+		App.scene.mac.LightsData.ambientLight.set(1, 1, 1);
+		App.scene.mac.mesh.setScale(3)
+	}
+	matrixEngine.objLoader.downloadMeshes(
+		{armor: "res/3d-objects/armor.obj", mac: "res/3d-objects/mac.obj"},
+		onLoadObj,
+		{swap: [0, 2]}
+	);
 };